requirements.txt
pandas
PyPDF2
python-dotenv
tavily-python